backend/requirements.txt
flask
flask-cors
python-dotenv

langchain==0.1.20
langchain-community==0.0.38
langchain-groq==0.1.5

sentence-transformers
faiss-cpu
requests
